evaluator/requirements.txt
langchain==0.3.27
langchain-community==0.3.29
translate-toolkit==3.16.0
langchain_openai==0.3.32
langchain_google_genai==2.1.10
langchain-community==0.3.29
llama-cpp-python==0.3.16
